requirements.txt
boto3==1.34.35
botocore==1.34.35
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
idna==3.6
jmespath==1.0.1
lxml==5.1.0
pillow==10.2.0
python-dateutil==2.8.2
python-docx==1.1.0
reportlab==4.0.9
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.0.7
